+++ History.md
@@ -1,4 +1,16 @@
-1.0.0 / HEAD
+1.1.1 / 2015-11-21
+==================
+
+* fix(KRR): prediction function is now working
+
+1.1.0 / 2015-11-19
+==================
+
+* Add Kernel Ridge Regression
+* SLR: add toString method
+* SLR: add computeX method
+
+1.0.0 / 2015-09-07
 ==================
 
 * first release